+++ changelog.md
@@ -4,6 +4,20 @@ All notable changes to the FolioBlocks project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2026-03-17
+
+### Fixed
+- Fixed bug that caused undefined variable warning in Grid Gallery
+
+## [1.2.2] - 2026-03-11
+
+### Added
+- Added responsive settings to Row Height slider in Justified Gallery Block.
+
+### Changed  
+- Updated Lightbox on Image Block to fill more of the viewport when exapnded. 
+- Increased maximum row height value on Justifed gallery block. 
+
 ## [1.2.1] - 2026-02-24
 
 ### Added